Bug 39877: Force commit id length to 10
The size of the commit id can vary, depending on the git repo.
For instance Jenkins had stored results using length of 9 (e9a07fe6a)
because it's a shallow clone.

But 9 is not enough for our repo:
% git rev-list --all | sort | uniq -w 9 -D
29dfcf49f28bb8e75ab4b71a2d68f194928d3de9
29dfcf49f68a2718c0d217921678cd1a0be38ab5

We need to force the length for both commands we use, so commits matched
during the comparison.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>