Bug 41112: Space is missing in report preview
This patch puts back a missing space between "Delete" and the fa
thrashcan. It is atm the lonely button without this space in the modal

No test plan needed. The change is too trivial.
Just check that lines before in the code have the desired space. The
missing space creates a weird sensation while looking at the window.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Brendan Lawlor <blawlor@clamsnet.org>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>