Bug 41088: (bug 38201 follow-up) Fix I18N for relationships
"Add new x", "Remove this x" and "There are no x created yet" were not
translatable correctly.

Test plan:
Add a new agreement, then a new title, and confirm that the 3 strings
are correctly displayed.
Then try to update .po files, translate the strings, and install the
translations and confirm that the strings are correctly translated when
using a translated interface.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
Signed-off-by: Matt Blenkinsop <matt.blenkinsop@openfifth.co.uk>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>